Add support for Mix configuration
Now that there is support for configuring your application via a
config/config.exs file, and the API for loading the config has been
exposed, we can merge that configuration with the generated sys.config
file.

Users can provide a sys.config file in the rel directory of their
project, and it will be merged over the top of the Mix configuration. If
no sys.config is provided, the Mix configuration will be merged over the
top of the default sys.config