Update boot script to be compatible with Solaris
Remove the use of the `local` operator, which is Bash specific and not
available in the Solaris version of `/bin/sh`.

To preserve the identifying fact that these are local variables, prefix
with a double underscore.

Changes inspired by
erlware/relx@cf5ed973dbfadd5edcc1d7d9e7ddcfcc202769df